refactor: extract budget history chart
