+++ frontend/src/budgets/components/budget-details-modal/BudgetHistoryChart.tsx
@@ -1,0 +1,171 @@
+import { useRef, type MouseEvent as ReactMouseEvent } from 'react'
+import {
+  Bar,
+  BarChart,
+  CartesianGrid,
+  ResponsiveContainer,
+  XAxis,
+  YAxis,
+} from 'recharts'
+import {
+  DeferredChartTooltipOverlay,
+  type DeferredChartTooltipOverlayHandle,
+} from '@/components/charts/DeferredChartTooltipOverlay'
+import {
+  getRechartsTooltipPoint,
+  getRechartsTooltipPointer,
+  type RechartsTooltipState,
+} from '@/components/charts/rechartsTooltip'
+import BudgetChartTooltip, { type BudgetChartPoint } from '@/budgets/components/budget-details-modal/BudgetChartTooltip'
+import { MODAL_SURFACE_TRANSITION_MS } from '@/budgets/constants'
+import {
+  BUDGET_CHART_HOVER_HIGHLIGHT_WIDTH,
+  BUDGET_CHART_LAYOUT,
+  getBudgetChartGuideMaxWidth,
+  type BudgetChartCategory,
+} from '@/budgets/utils/budgetDetails'
+
+const CHART_INITIAL_DIMENSION = { width: 1, height: 192 }
+
+type BudgetHistoryChartProps = {
+  chartData: BudgetChartPoint[]
+  chartCategories: BudgetChartCategory[]
+  currency: string
+  loading: boolean
+  error: boolean
+}
+
+function getBudgetChartTooltipKey(point: BudgetChartPoint) {
+  return point.label
+}
+
+/**
+ * Renders the budget details historical utilization chart and its deferred tooltip
+ */
+export default function BudgetHistoryChart({
+  chartData,
+  chartCategories,
+  currency,
+  loading,
+  error,
+}: BudgetHistoryChartProps) {
+  const chartRef = useRef<HTMLDivElement>(null)
+  const tooltipRef = useRef<DeferredChartTooltipOverlayHandle<BudgetChartPoint>>(null)
+  const showStackedCategoryChart = chartCategories.length > 1
+
+  /**
+   * Resolves the hovered Recharts bar and forwards it to the shared deferred tooltip overlay
+   */
+  function showTooltip(
+    state: RechartsTooltipState<BudgetChartPoint>,
+    event: ReactMouseEvent<SVGGraphicsElement>,
+  ) {
+    const point = getRechartsTooltipPoint({
+      state,
+      data: chartData,
+      resolveLabel: (label) => chartData.find((item) => item.label === label),
+    })
+    const pointer = getRechartsTooltipPointer(state, event)
+
+    if (!point) {
+      tooltipRef.current?.show(null, pointer)
+      return
+    }
+
+    tooltipRef.current?.show(point, pointer)
+  }
+
+  const hideTooltip = () => tooltipRef.current?.hide()
+
+  return (
+    <div
+      ref={chartRef}
+      className="relative h-48 min-[750px]:h-80"
+      onMouseLeave={hideTooltip}
+    >
+      {loading ? (
+        <div
+          className="flex h-full items-center justify-center rounded-xl"
+          style={{ background: 'var(--app-bg)' }}
+        >
+          <div className="app-spinner" />
+        </div>
+      ) : error ? (
+        <div
+          className="flex h-full items-center justify-center rounded-xl text-sm"
+          style={{ background: 'var(--app-bg)', color: 'var(--app-negative)' }}
+        >
+          Utilization history could not load.
+        </div>
+      ) : chartData.length > 0 ? (
+        <>
+          <ResponsiveContainer width="100%" height="100%" initialDimension={CHART_INITIAL_DIMENSION}>
+            <BarChart
+              data={chartData}
+              margin={BUDGET_CHART_LAYOUT.margin}
+              onMouseMove={(state, event) => showTooltip(state, event)}
+              onMouseLeave={hideTooltip}
+            >
+              <CartesianGrid stroke="var(--app-border)" vertical={false} />
+              <XAxis
+                dataKey="label"
+                tickLine={false}
+                axisLine={false}
+                tick={{ fill: 'var(--app-text-subtle)', fontSize: 13 }}
+              />
+              <YAxis
+                tickLine={false}
+                axisLine={false}
+                domain={[0, (dataMax: number) => Math.max(100, Math.ceil(dataMax / 25) * 25)]}
+                tick={{ fill: 'var(--app-text-subtle)', fontSize: 12 }}
+                tickFormatter={(value) => `${Number(value)}%`}
+                width={BUDGET_CHART_LAYOUT.yAxisWidth}
+              />
+              {showStackedCategoryChart ? chartCategories.map((category, index) => (
+                <Bar
+                  key={category.id}
+                  dataKey={category.dataKey}
+                  stackId="category-spending"
+                  fill={category.color}
+                  radius={index === chartCategories.length - 1 ? [4, 4, 0, 0] : [0, 0, 0, 0]}
+                  barSize={28}
+                  animationBegin={MODAL_SURFACE_TRANSITION_MS}
+                />
+              )) : (
+                <Bar
+                  dataKey="utilizationPct"
+                  fill="var(--app-accent)"
+                  radius={[4, 4, 0, 0]}
+                  barSize={28}
+                  animationBegin={MODAL_SURFACE_TRANSITION_MS}
+                />
+              )}
+            </BarChart>
+          </ResponsiveContainer>
+          <DeferredChartTooltipOverlay
+            ref={tooltipRef}
+            chartRef={chartRef}
+            className="min-w-44"
+            guideVariant="bar"
+            guideWidth={BUDGET_CHART_HOVER_HIGHLIGHT_WIDTH}
+            guideMaxWidth={(chartWidth) => getBudgetChartGuideMaxWidth(chartWidth, chartData.length)}
+            getKey={getBudgetChartTooltipKey}
+            renderContent={(point) => (
+              <BudgetChartTooltip
+                point={point}
+                currency={currency}
+              />
+            )}
+          />
+        </>
+      ) : (
+        <div
+          className="flex h-full items-center justify-center rounded-xl text-sm"
+          style={{ background: 'var(--app-bg)', color: 'var(--app-text-subtle)' }}
+        >
+          No utilization history yet.
+        </div>
+      )}
+    </div>
+  )
+}
